+++ Cargo.toml
@@ -6,12 +6,12 @@ edition = "2024"
 
 [features]
 default = ["shell"]
-all-inputs = ["shell", "ros1", "ros2", "metrics", "mqtt"]
+all-inputs = ["shell", "ros1", "ros2", "metrics", "iot"]
 ros1 = ["dep:rosrust"]
 ros2 = ["dep:rclrs", "dep:ros2_message"]
 shell = []
 metrics = ["dep:systemstat"]
-mqtt = ["dep:rumqttc", "dep:rustls"]
+iot = ["dep:rumqttc", "dep:rustls"]
 ci = []
