+++ examples/mqtt_config.toml
@@ -18,8 +18,6 @@ prefix = ""
 broker = "mqtts://broker.example.com:8883"
 client_id = "reduct-bridge"
 version = "v5"
-
-topics = ["factory/+/telemetry", "factory/+/events"]
 qos = 1
 
 username = "bridge"
@@ -27,16 +25,28 @@ password = "${MQTT_PASSWORD}"
 
 entry_prefix = "/mqtt"
 
-[inputs.mqtt.main.labels]
-device = "$.device_id"
-site = "$.site"
-static_source = "mqtt"
+[[inputs.mqtt.main.topics]]
+name = "factory/+/telemetry"
+entry_name = "telemetry"
+content_type = "application/json"
 
-[inputs.mqtt.main.property_labels]
-reply = "response_topic"
-mime = "content_type"
-correlation = "correlation_data"
-tenant = "user.tenant"
+[[inputs.mqtt.main.topics]]
+name = "factory/+/events"
+
+# Optional label rules (default = []):
+# 1) Dynamic field label:
+#    { field = "device_id", label = "device" }
+# 2) Static labels:
+#    { static = { source = "mqtt", site = "lab" } }
+# 3) MQTT v5 property label:
+#    { property = "user.tenant", label = "tenant" }
+labels = [
+  { field = "device_id", label = "device" },
+  { field = "site", label = "site" },
+  { static = { source = "mqtt" } },
+  { property = "content_type", label = "mime" },
+  { property = "user.tenant", label = "tenant" }
+]
 
 
 # MQTT v3 input configuration.
@@ -44,8 +54,6 @@ tenant = "user.tenant"
 broker = "mqtt://broker.example.com:1883"
 client_id = "reduct-bridge-legacy"
 version = "v3"
-
-topics = ["legacy/+/data"]
 qos = 0
 
 username = "legacy_user"
@@ -53,9 +61,15 @@ password = "${LEGACY_MQTT_PASSWORD}"
 
 entry_prefix = "/mqtt"
 
-[inputs.mqtt.legacy.labels]
-line = "$.line"
-static_source = "mqtt-v3"
+[[inputs.mqtt.legacy.topics]]
+name = "legacy/+/data"
+entry_name = "legacy"
+content_type = "application/json"
+
+labels = [
+  { field = "line", label = "line" },
+  { static = { source = "mqtt-v3" } }
+]
 
 
 # Pipelines
