+++ src/input/iot/mod.rs
@@ -1,0 +1,1 @@
+pub mod mqtt;
