+++ src/input/mqtt/README.md
@@ -3,12 +3,11 @@
 The MQTT input subscribes to one or more MQTT topics and forwards messages into a pipeline.
 Both `mqtt://` and `mqtts://` brokers are supported.
 
-Payload label rules use the following convention:
+MQTT label rules follow the same ordered model as the ROS inputs:
 
-- values starting with `$.` are treated as JSON field paths inside the MQTT payload
-- all other values are treated as static label values
-
-MQTT v5 adds `property_labels`, which map selected MQTT v5 publish properties into record labels.
+- static labels are applied first
+- payload field labels are applied after static labels and can override them
+- MQTT v5 property labels are also applied in order and can override earlier labels
 
 ## Configuration
 
@@ -17,8 +16,6 @@ MQTT v5 adds `property_labels`, which map selected MQTT v5 publish properties in
 broker = "mqtts://broker.example.com:8883"
 client_id = "reduct-bridge"
 version = "v5"
-
-topics = ["factory/+/telemetry", "factory/+/events"]
 qos = 1
 
 username = "bridge"
@@ -26,17 +23,24 @@ password = "${MQTT_PASSWORD}"
 
 entry_prefix = "/mqtt"
 
-[inputs.mqtt.main.labels]
-device = "$.device_id"
-site = "$.site"
-static_source = "mqtt"
+[[inputs.mqtt.main.topics]]
+name = "factory/+/telemetry"
+entry_name = "telemetry"
+content_type = "application/json"
 
-[inputs.mqtt.main.property_labels]
-reply = "response_topic"
-mime = "content_type"
-correlation = "correlation_data"
-tenant = "user.tenant"
+[[inputs.mqtt.main.topics]]
+name = "factory/+/events"
+
+labels = [
+  { field = "device_id", label = "device" },
+  { field = "site", label = "site" },
+  { static = { source = "mqtt" } },
+  { property = "content_type", label = "mime" },
+  { property = "user.tenant", label = "tenant" }
+]
 ```
+
+`entry_name` is optional per topic. `content_type` is used as the default record content type when an MQTT v5 publish does not provide a `content_type` property; for MQTT v3 it is used directly.
 
 ### MQTT v3 example
 
@@ -45,8 +49,6 @@ tenant = "user.tenant"
 broker = "mqtt://broker.example.com:1883"
 client_id = "reduct-bridge-legacy"
 version = "v3"
-
-topics = ["legacy/+/data"]
 qos = 0
 
 username = "legacy_user"
@@ -54,7 +56,13 @@ password = "${LEGACY_MQTT_PASSWORD}"
 
 entry_prefix = "/mqtt"
 
-[inputs.mqtt.legacy.labels]
-line = "$.line"
-static_source = "mqtt-v3"
+[[inputs.mqtt.legacy.topics]]
+name = "legacy/+/data"
+entry_name = "legacy"
+content_type = "application/json"
+
+labels = [
+  { field = "line", label = "line" },
+  { static = { source = "mqtt-v3" } }
+]
 ```
